Rather than setting an intrinsic type's kind using an integer literal, such as real(8) or integer(kind=4), set kinds using parameters in the intrinsic module iso_fortran_env such as real64 and int32.

+++ src/io_hdf5/hdf5_output_particle.f90
@@ -1453,7 +1453,7 @@ INTEGER(KIND=IK),INTENT(OUT) :: globnPart(6)
 !-----------------------------------------------------------------------------------------------------------------------------------
 ! LOCAL VARIABLES
 #if USE_MPI
-INTEGER(KIND=8)              :: locnPart8,locnPart8Recv,globnPart8 ! always integer KIND=8
+INTEGER(KIND=i8)             :: locnPart8,locnPart8Recv,globnPart8 ! always integer KIND=8
 INTEGER(KIND=IK)             :: SimNumSpecMin,SimNumSpecMax
 #else
 CHARACTER(LEN=255) :: dummy_char
